move user initialization to App
